update docs with tokenizers
