+++ client/src/bastionlab/polars/_utils.py
@@ -1,15 +1,116 @@
-from typing import Iterator, List
+from typing import Iterator, List, Union, TYPE_CHECKING
 import polars as pl
 import io
 from ..pb.bastionlab_polars_pb2 import SendChunk
 from .policy import Policy
 from serde.json import to_json
+from dataclasses import dataclass
+from torch.jit import ScriptFunction
+import base64
+import json
+from serde import serde, InternalTagging, field
+from serde.json import to_json
+
+if TYPE_CHECKING:
+    from .client import BastionLabPolars
 
 CHUNK_SIZE = 32 * 1024
 
 # TODO PERF: Do a PR on polars/pypolars to add the streaming IPC (apache flight) format to the python interface
 # right now, there is only the file format which requires random access
 # which means, we have to do a full copy to a buffer and we cannot parse it as we go
+
+
+class CompositePlanSegment:
+    """
+    Composite plan segment class which handles segment plans that have not been implemented
+    """
+
+
+@dataclass
+@serde
+class EntryPointPlanSegment(CompositePlanSegment):
+    """
+    Composite plan segment class responsible for new entry points
+    """
+
+    identifier: str
+
+
+@dataclass
+@serde
+class PolarsPlanSegment(CompositePlanSegment):
+    """
+    Composite plan segment class responsible for Polars queries
+    """
+
+    # HACK: when getting using the schema attribute, polars returns
+    #  the proper error messages (polars.NotFoundError etc) when it is invalid.
+    #  This is not the case for write_json(), which returns a confusing error
+    #  message. So, we get the schema beforehand :)
+    plan: pl.LazyFrame = field(
+        serializer=lambda val: val.schema and json.loads(val.write_json()),
+        deserializer=lambda _: None,
+    )
+
+
+@dataclass
+@serde
+class UdfPlanSegment(CompositePlanSegment):
+    """
+    Composite plan segment class responsible for user defined functions
+    """
+
+    columns: List[str]
+    udf: ScriptFunction = field(
+        serializer=lambda val: base64.b64encode(val.save_to_buffer()).decode("ascii"),
+        deserializer=lambda _: None,
+    )
+
+
+@dataclass
+@serde
+class StackPlanSegment(CompositePlanSegment):
+    """
+    Composite plan segment class responsible for vstack function
+    """
+
+
+@dataclass
+@serde
+class RowCountSegment(CompositePlanSegment):
+    """
+    Composite plan segment class responsible for with_row_count function
+    """
+
+    row: str
+
+
+@dataclass
+@serde(tagging=InternalTagging("type"))
+class PlanSegments:
+    segments: List[
+        Union[
+            PolarsPlanSegment,
+            UdfPlanSegment,
+            EntryPointPlanSegment,
+            StackPlanSegment,
+            RowCountSegment,
+        ]
+    ]
+
+
+@dataclass
+class UdfTransformerPlanSegment(CompositePlanSegment):
+    """
+    Accepts a UDF for row-wise DataFrame transformation.
+    """
+
+    _name: str
+    _columns: List[str]
+
+    def serialize(self) -> str:
+        pass
 
 
 def serialize_dataframe(
@@ -70,3 +171,13 @@ def deserialize_dataframe(chunks: Iterator[bytes]) -> pl.DataFrame:
     df = pl.read_ipc(buf)
 
     return df
+
+
+@dataclass
+class Metadata:
+    """
+    A class containing metadata related to your dataframe
+    """
+
+    _polars_client: "BastionLabPolars"
+    _prev_segments: List[CompositePlanSegment] = field(default_factory=list)
