+++ client/src/bastionlab/tokenizers/__init__.py
@@ -1,1 +1,9 @@
 from .client import BastionLabTokenizers
+from .remote_tokenizers import RemoteTokenizer
+
+__pdoc__ = {}
+
+__pdoc__["BastionLabTokenizers.__init__"] = False
+__pdoc__["RemoteTokenizer.__init__"] = False
+
+__all__ = ["BastionLabTokenizers", "RemoteTokenizer"]
